backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Cache
redis==5.0.1

# ML/AI
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
torch==2.1.1
transformers==4.36.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1

# Monitoring
prometheus-client==0.19.0
structlog==23.2.0
openai==1.3.7
